+++ frontend/src/lib/components/sidebar/SessionList.svelte
@@ -572,7 +572,7 @@
   />
 {/if}
 
-<div class="session-list flex h-full flex-col" data-testid="session-list">
+<div class="session-list flex flex-1 min-h-0 flex-col" data-testid="session-list">
   <TagFilterPanel
     tags={tagsStore.all}
     selectedProjectIds={tagsStore.selectedProjectIds}